Allow for values outside of the range of [1..length]
Populate the array of values with numbers from 0 to 2^10-1.
Maintain a separate array of "indices" whose sort order matches the sort
order of the values. This allows for a nice even triangle when sorted,
whereas using the values directly would provide an uneven slope.

Update all of the swap and insert functions to take advantage of
CoffeeScript's array destructuring syntax. This means no more tmp
variables.

Reimplement shuffle as a Fisher-Yates shuffle.