Don't blow away animation state on highlighting
The calls to VA.highlight() and VA.persistHighlight() were pushing an
animation frame that omitted all of the normal animation state,
including locals. This caused, for example, radix sort with a
VA.locals.bit = bit line to keep losing it's display of "bit" as it
cycled downwards, and only restoring it once it performed another swap.

Fix this by replacing @animationQueue.push() with @animationQueuePush().